Революция или эволюция? - правки
